remove int8 bug

+++ src/contracts/mobcoin.sol
@@ -6,7 +6,7 @@ contract MobCoin {
     // Public variables of the token
     string public name;
     string public symbol;
-    uint8 public decimals = 18;
+    uint16 public decimals = 18;
     // 18 decimals is the strongly suggested default, avoid changing it
     uint256 public totalSupply;
     address owner;
